move ReadOnlyRow component to Estimate folder
